+++ src/views/deal/DealChuzu.vue
@@ -48,7 +48,16 @@
                 <el-form-item label="付款方式：">
                   <span>{{ props.row.payway }}</span>
                 </el-form-item>
-                <el-form-item label="合同录入：">
+                <el-form-item label="应收佣金：">
+                  <span>{{ props.row.price3 }}</span>
+                </el-form-item>
+                <el-form-item label="实收佣金：">
+                  <span>{{ props.row.price4 }}</span>
+                </el-form-item>
+                <el-form-item label="是否退佣：">
+                  <span>{{ props.row.tuiyong }}</span>
+                </el-form-item>
+                <el-form-item label="合同录入(文字)：">
                   <span>{{ props.row.details }}</span>
                 </el-form-item>
 
@@ -62,6 +71,10 @@
           <el-table-column
             label="坐落位置"
             prop="place">
+          </el-table-column>
+          <el-table-column
+            label="成交业务员"
+            prop="upname">
           </el-table-column>
           <el-table-column
             label="甲方"
@@ -80,7 +93,7 @@
             prop="endDate">
           </el-table-column>
           <el-table-column
-            width="100">
+            width="200">
             <template slot-scope="scope">
               <el-button
                 @click.native.prevent="showChuzudeal(scope.row,scope.$index)"
@@ -94,125 +107,44 @@
                 size="small">
                 删除
               </el-button>
+              <el-button
+                @click.native.prevent="getImg(scope.row,scope.$index)"
+                type="text"
+                size="small">
+                合同录入（图片）
+              </el-button>
             </template>
           </el-table-column>
 
         </el-table>
-        <!--
-        <el-table
-          :data="emps"
-          stripe
-          border
-          style="width: 100%">
-          <el-table-column
-            fixed
-            type="selection"
-            width="55">
-          </el-table-column>
-          <el-table-column
-          fixed
-          prop="workid"
-          label="编号"
-          width="100">
-        </el-table-column>
-          <el-table-column
-            fixed
-            prop="name"
-            label="小区名称"
-            width="120">
-          </el-table-column>
-          <el-table-column
-            fixed
-          prop="idcard"
-          label="房号"
-          width="130">
-        </el-table-column>
-          <el-table-column
-            fixed
-            prop="gender"
-            label="是否满2"
-            width="60">
-          </el-table-column>
-          <el-table-column
-            fixed
-            prop="car"
-            label="车库/车位"
-            width="88">
-          </el-table-column>
 
-          <el-table-column
-            fixed
-            prop="email"
-            label="面积"
-            width="70">
-          </el-table-column>
-          <el-table-column
-            fixed
-            prop="phone"
-            label="接待人"
-            width="100">
-          </el-table-column>
-          <el-table-column
-            fixed
-            prop="reno"
-            label="装修/毛坯"
-            width="70">
-          </el-table-column>
+        <el-dialog
+          :title="title"
+          :visible.sync="dialogUploadVisible"
+          width="70%"
+          @close='closeDialog'>
+          <el-upload
+            action="/employee/basic/img/"
+            accept="image/png, image/jpeg"
+            list-type="picture-card"
+            :file-list='this.imgSrcs'
+            :before-upload="beforeUploadPicture"
+            :on-preview="handlePictureCardPreview"
+            :on-progress="uploadProgress"
+            :on-success="uploadSuccess"
+            :on-change="uploadLimit"
+            :on-error="uploadError"
+            :before-remove="beforeRemove"
+            :on-remove="handleRemove"
+            :class="{disabled:uploadDisabled}"
+            :show-file-list="true">
+            <i class="el-icon-plus"></i>
+          </el-upload>
+          <el-dialog :visible.sync="dialogImageVisible">
+            <el-image :lazy="true"  width="100%" :src="dialogImageUrl" alt=""/>
+          </el-dialog>
+        </el-dialog>
 
-          <el-table-column
-            fixed
-            prop="price"
-            label="价格"
-            width="70">
-          </el-table-column>
-          <el-table-column
-            fixed
-            prop="intr"
-            label="上门/介绍"
-            width="70">
-          </el-table-column>
-          <el-table-column
-            fixed
-          prop="looktime"
-          label="看房时间"
-          width="120">
-        </el-table-column>
-          <el-table-column
-            fixed
-            prop="phone2"
-            label="联系方式"
-            width="140">
-          </el-table-column>
-          <el-table-column
-            fixed="right"
-            label="操作"
-            width="100">
-            <template slot-scope="scope">
-              <el-button
-                @click.native.prevent="showEmp(scope.row,scope.$index)"
-                type="text"
-                size="small">
-                编辑
-              </el-button>
-              <el-button
-                @click.native.prevent="deleteEmp(scope.row,scope.$index)"
-                type="text"
-                size="small">
-                删除
-              </el-button>
-            </template>
-          </el-table-column>
-        </el-table>
-        <div style="display: flex;justify-content: flex-end">
-          <el-pagination
-            background
-            @current-change="currentChange"
-            @size-change="sizeChange"
-            layout="sizes, prev, pager, next, jumper, ->, total, slot"
-            :total="total">
-          </el-pagination>
-        </div>
-        -->
       </div>
       <el-dialog
         :title="title"
@@ -280,6 +212,23 @@
                 </el-form-item>
               </el-col>
             </el-row>
+            <el-row :gutter="1">
+              <el-col :span="8">
+                <el-form-item label="应收佣金：" prop="price3">
+                  <el-input  style="width:150px" placeholder="请输入应收佣金" v-model="chuzudeal.price3"></el-input>
+                </el-form-item>
+              </el-col>
+              <el-col :span="8">
+                <el-form-item label="实收佣金：" prop="price4">
+                  <el-input  style="width:150px" placeholder="请输入实收佣金" v-model="chuzudeal.price4"></el-input>
+                </el-form-item>
+              </el-col>
+              <el-col :span="8">
+                <el-form-item label="是否退佣：" prop="tuiyong">
+                  <el-input  style="width:150px" placeholder="请输入" v-model="chuzudeal.tuiyong"></el-input>
+                </el-form-item>
+              </el-col>
+            </el-row>
             <el-row :gutter="20">
               <el-col :span="8">
                 <el-form-item label="编号：" prop="workid">
@@ -317,64 +266,7 @@
                 placeholder="请输入内容" v-model="chuzudeal.details"
               ></el-input>
             </el-row>
-         <!--   <el-row :gutter="20">
-              <el-col :span="8">
-                <el-form-item label="所属部门：" prop="departmentid">
-                  <el-popover
-                    placement="right"
-                    title="请选择部门"
-                    width="200"
-                    trigger="manual"
-                    v-model="visible">
-                    <el-tree default-expand-all :data="allDeps" :props="defaultProps" @node-click="handleNodeClick"></el-tree>
-                    <div slot="reference" style="width: 200px;display: inline-flex;font-size: 13px;border: 1px solid #dedede;height: 35px;border-radius: 5px;
-                  cursor: pointer;align-items: center;padding-left: 8px" @click="showDepView">{{inputDepName}}</div>
-                  </el-popover>
-                </el-form-item>
-              </el-col>
-              <el-col :span="8">
-                <el-form-item label="职称：" prop="joblevelid">
-                  <el-select v-model="emp.joblevelid" placeholder="请选择">
-                    <el-option
-                      v-for="item in joblevels"
-                      :key="item.id"
-                      :label="item.name"
-                      :value="item.id">
-                    </el-option>
-                  </el-select>
-                </el-form-item>
-              </el-col>
-              <el-col :span="8">
-                <el-form-item label="职位：" prop="posid">
-                  <el-select v-model="emp.posid" placeholder="请选择">
-                    <el-option
-                      v-for="item in positions"
-                      :key="item.id"
-                      :label="item.name"
-                      :value="item.id">
-                    </el-option>
-                  </el-select>
-                </el-form-item>
-              </el-col>
-            </el-row>
-            <el-row :gutter="20">
-              <el-col :span="8">
-                <el-form-item label="入职日期：" prop="begindate">
-                  <el-date-picker
-                    v-model="emp.begindate"
-                    style="width: 200px"
-                    type="date"
-                    value-format="yyyy-MM-dd"
-                    placeholder="选择日期">
-                  </el-date-picker>
-                </el-form-item>
-              </el-col>
-              <el-col :span="8">
-                <el-form-item label="工号：" prop="workid">
-                  <el-input  style="width:150px" v-model="emp.workid" disabled></el-input>
-                </el-form-item>
-              </el-col>
-            </el-row>-->
+
           </el-form>
         </div>
         <span slot="footer" class="dialog-footer">
@@ -405,6 +297,19 @@
     name: 'ChuzudealBasic',
     data(){
       return{
+        imgpath:{
+          workid:"",
+          imagepath:"",
+        },
+        rowa:"",
+        upnames:"",
+        urls:[],
+        user:JSON.parse(window.sessionStorage.getItem("user")),
+        imgSrcs:[],
+        uploadDisabled:false,
+        dialogImageUrl: [],
+        dialogImageVisible:false,
+        dialogUploadVisible:false,
         title:'',
         inputDepName:'',
         allDeps:[],
@@ -434,8 +339,12 @@
           workid:"",
           price1:"",
           price2:"",
+          price3:"",
+          price4:"",
+          upname:"",
           payway:"",
           details:"",
+          tuiyong:"",
         },
         defaultProps: {
           children: 'children',
@@ -447,6 +356,8 @@
           yifang:[{required:true,message:'请输入乙方',trigger:'blur'}],
           price1:[{required:true,message:'请输入租金',trigger:'blur'}],
           payway:[{required:true,message:'请输入付款方式',trigger:'blur'}],
+          price3:[{required:true,message:'请输入应收佣金',trigger:'blur'}],
+          price4:[{required:true,message:'请输入实收佣金',trigger:'blur'}],
         }
 
       }
@@ -458,6 +369,87 @@
       // this.initData();
     },
     methods:{
+      getImg(row){
+        console.log(row);
+        this.rowa = row.workid;
+        this.upnames = row.upname;
+        this.getRequest("/employee/basic/getImg/?workid="+this.rowa).then(res=> {
+
+          console.log(res);
+          this.imgpaths = res.data;
+          console.log(this.imgpaths);
+          for(let i=0;i<this.imgpaths.length;i++){
+            const imgSrc = {
+              name:"",
+              url:"",
+            }
+            imgSrc.name = this.imgpaths[i].id;
+            imgSrc.url= "../../static"+this.imgpaths[i].imagepath;
+            this.imgSrcs.push(imgSrc);
+          }
+
+          console.log(this.imgSrcs);
+          this.dialogUploadVisible = true;
+        })
+      },
+      closeDialog(){
+        this.imgSrcs=[];
+        this.dialogUploadVisible=false;
+
+      },
+      handleRemove(file, fileList) {
+        if(this.user.name==this.upnames||this.user.roles[0].id==3||this.user.roles[0].id==6){
+          this.deleteRequest("/employee/basic/deleteImg/"+file.name)
+          console.log(file, fileList);
+        }else{
+          this.$confirm("非上传者无权删除图片，刷新后再次出现");
+        }
+
+      },
+      uploadSuccess(response, file, fileList) {
+        console.log(response);
+        console.log(this.rowa);
+        if(this.user.name==this.upnames||this.user.roles[0].id==3||this.user.roles[0].id==6){
+          this.imgpath.imagepath = response;
+          this.imgpath.workid = this.rowa;
+          this.postRequest("/employee/basic/addImg",this.imgpath)
+        }else{
+          this.$confirm("非上传者无权上传图片，刷新后恢复");
+        }
+      }
+      ,
+      uploadLimit(file,fileList){
+        console.log(this.upnames);
+        console.log(this.user.name);
+        if(this.upnames!=this.user.name){
+          this.uploadDisabled = true;
+        }
+      },
+      beforeRemove(file, fileList) {
+
+        return this.$confirm(`确定移除 ${ file.name }？`);
+
+      },
+
+      beforeUploadPicture(file) {
+        if(file.size > 10*1024*1024){
+          this.$message.error("上传图片不能大于10M");
+          return false;
+        }
+      },
+      // handleRemove(file, fileList) {
+      //   console.log(file, fileList);
+      // },
+      handlePictureCardPreview(file) {
+        this.dialogImageUrl = file.url;
+        this.dialogImageVisible = true;
+      },
+      uploadProgress(event,file, fileList){
+
+      },
+      uploadError(err, file, fileList) {
+        this.$message.error("上传出错");
+      },
       emptyChuzudeal(){
         this.chuzudeal={
           place:"",
@@ -475,8 +467,12 @@
           workid:"",
           price1:"",
           price2:"",
+          price3:"",
+          price4:"",
+          upname:"",
           payway:"",
           details:"",
+          tuiyong:"",
         }
       },
  doAddChuzudeal(){
@@ -492,6 +488,7 @@
         }else {
           this.$refs['chuzudealForm'].validate(valid=>{
             if (valid){
+              this.chuzudeal.upname=this.user.name;
               this.postRequest("/deal/chuzu/",this.chuzudeal).then(res=>{
                 this.dialogVisible = false;
                 this.initChuzudeals();
@@ -533,31 +530,43 @@
         this.dialogVisible=true;
       },
       deleteChuzudeal(row){
-        this.$confirm('此操作将永久删除【'+row.place+'】房源, 是否继续?', '提示', {
-          confirmButtonText: '确定',
-          cancelButtonText: '取消',
-          type: 'warning'
-        }).then(() => {
-          this.deleteRequest("/deal/chuzu/"+row.id).then(res=>{
-            if (res){
-              this.initChuzudeals();
-            }
-          })
-        }).catch(() => {
-          this.$message({
-            type: 'info',
-            message: '已取消删除'
+        console.log(this.user.name);
+        console.log(row.upname);
+        if (row.upname != this.user.name && this.user.name != "管理员" && this.user.name != "系统管理员") {
+          this.$confirm("非上传者无法删除");
+        } else {
+          this.$confirm('此操作将永久删除【' + row.place + '】房源, 是否继续?', '提示', {
+            confirmButtonText: '确定',
+            cancelButtonText: '取消',
+            type: 'warning'
+          }).then(() => {
+            this.deleteRequest("/deal/chuzu/" + row.id).then(res => {
+              if (res) {
+                this.initChuzudeals();
+              }
+            })
+          }).catch(() => {
+            this.$message({
+              type: 'info',
+              message: '已取消删除'
+            });
           });
-        });
+        }
       },
       showChuzudeal(row){
+        console.log(row.upname);
+        console.log(this.user.name);
+        if (row.upname == this.user.name || this.user.name == "管理员" || this.user.name == "系统管理员") {
         this.title='编辑房源信息';
         if(this.$refs['chuzudealForm']!=undefined){
           this.$refs['chuzudealForm'].resetFields()
         }
         this.chuzudeal=row;
         this.dialogVisible=true;
-      }
+      }else {
+          this.$confirm("非上传者无法编辑");
+        }
+    }
     }
   }
 </script>
